+++ src/extension.ts
@@ -1,7 +1,6 @@
 'use strict';
 
 import vscode from 'vscode';
-import lockFile from 'lockfile';
 
 import { eventHandler } from "./events/event_handler";
 import { setupCodeSync, showConnectRepoView, showLogIn } from "./utils/setup_utils";
@@ -20,8 +19,6 @@ import {
 	openSyncIgnoreHandler
 } from './handlers/commands_handler';
 import { putLogEvent } from "./logger";
-import { generateSettings } from './settings';
-import { CodeSyncState, CODESYNC_STATES } from './utils/state_utils';
 
 export async function activate(context: vscode.ExtensionContext) {
 	try {
@@ -108,22 +105,5 @@ export async function activate(context: vscode.ExtensionContext) {
 }
 
 export function deactivate(context: vscode.ExtensionContext) {
-	const settings = generateSettings();
-	// If locks were acquired by current instance, release the locks
-	const canRunPopulateBuffer = CodeSyncState.get(CODESYNC_STATES.POPULATE_BUFFER_LOCK_ACQUIRED);
-	const canRunBufferHandler = CodeSyncState.get(CODESYNC_STATES.DIFFS_SEND_LOCK_ACQUIRED);
-	if (canRunPopulateBuffer) {
-		try {
-			lockFile.unlockSync(settings.POPULATE_BUFFER_LOCK_FILE);
-		} catch (e) {
-			// 
-		}
-	}
-	if (canRunBufferHandler) {
-		try {
-			lockFile.unlockSync(settings.DIFFS_SEND_LOCK_FILE);
-		} catch (e) {
-			// 
-		}
-	}
+	// Empty body for now
 }