+++ tests/test_codesyncd/codesyncd.test.js
@@ -1,6 +1,6 @@
 import fs from "fs";
 import path from "path";
-import lockFile from "lockfile";
+import lockFile from "proper-lockfile";
 import vscode from "vscode";
 import untildify from "untildify";
 import {generateSettings} from "../../src/settings";
@@ -18,6 +18,7 @@ import {
 import {createSystemDirectories} from "../../src/utils/setup_utils";
 import {STATUS_BAR_MSGS, COMMAND, SYNCIGNORE} from "../../src/constants";
 import {statusBarMsgs} from "../../src/codesyncd/utils";
+import { LockUtils } from "../../src/utils/lock_utils";
 
 describe("codesyncd: locks", () => {
     const baseRepoPath = randomBaseRepoPath();
@@ -34,13 +35,12 @@ describe("codesyncd: locks", () => {
     });
 
     afterEach(() => {
-        lockFile.unlockSync(settings.POPULATE_BUFFER_LOCK_FILE);
-        lockFile.unlockSync(settings.DIFFS_SEND_LOCK_FILE);
         fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
 
     test("acquirePopulateBufferLock", () => {
-        acquirePopulateBufferLock();
+        const lockUtils = new LockUtils();
+        lockUtils.acquirePopulateBufferLock();
         expect(lockFile.checkSync(settings.POPULATE_BUFFER_LOCK_FILE)).toBe(true);
         expect(CodeSyncState.get(CODESYNC_STATES.POPULATE_BUFFER_LOCK_ACQUIRED)).toBe(true);
 		expect(lockFile.checkSync(settings.DIFFS_SEND_LOCK_FILE)).toBe(false);
@@ -48,7 +48,8 @@ describe("codesyncd: locks", () => {
     });
 
 	test("acquireSendDiffsLock", () => {
-        acquireSendDiffsLock();
+        const lockUtils = new LockUtils();
+        lockUtils.acquireSendDiffsLock();
 		expect(lockFile.checkSync(settings.DIFFS_SEND_LOCK_FILE)).toBe(true);
 		expect(CodeSyncState.get(CODESYNC_STATES.DIFFS_SEND_LOCK_ACQUIRED)).toBe(true);
         expect(lockFile.checkSync(settings.POPULATE_BUFFER_LOCK_FILE)).toBe(false);
@@ -80,8 +81,6 @@ describe("codesyncd: recallDaemon", () => {
     });
 
     afterEach(() => {
-        lockFile.unlockSync(settings.POPULATE_BUFFER_LOCK_FILE);
-        lockFile.unlockSync(settings.DIFFS_SEND_LOCK_FILE);
         fs.rmSync(repoPath, { recursive: true, force: true });
         fs.rmSync(baseRepoPath, { recursive: true, force: true });
     });
@@ -161,7 +160,8 @@ describe("codesyncd: recallDaemon", () => {
     });
 
     test("with lock acquired for populateBuffer", () => {
-        acquirePopulateBufferLock();
+        const lockUtils = new LockUtils();
+        lockUtils.acquirePopulateBufferLock();
         recallDaemon(statusBarItem);
         expect(lockFile.checkSync(settings.POPULATE_BUFFER_LOCK_FILE)).toBe(true);
 		expect(lockFile.checkSync(settings.DIFFS_SEND_LOCK_FILE)).toBe(true);
@@ -170,7 +170,8 @@ describe("codesyncd: recallDaemon", () => {
     });
 
     test("with lock acquired for diffsSend", () => {
-        acquireSendDiffsLock();
+        const lockUtils = new LockUtils();
+        lockUtils.acquireSendDiffsLock();
         recallDaemon(statusBarItem);
         expect(lockFile.checkSync(settings.POPULATE_BUFFER_LOCK_FILE)).toBe(true);
 		expect(lockFile.checkSync(settings.DIFFS_SEND_LOCK_FILE)).toBe(true);
@@ -180,7 +181,8 @@ describe("codesyncd: recallDaemon", () => {
 
     test("with diffsSendLock acquried by other instance", () => {
         lockFile.lockSync(settings.DIFFS_SEND_LOCK_FILE);
-        acquirePopulateBufferLock();
+        const lockUtils = new LockUtils();
+        lockUtils.acquirePopulateBufferLock();
         recallDaemon(statusBarItem);
         expect(CodeSyncState.get(CODESYNC_STATES.POPULATE_BUFFER_LOCK_ACQUIRED)).toBe(true);
 		expect(CodeSyncState.get(CODESYNC_STATES.DIFFS_SEND_LOCK_ACQUIRED)).toBe(false);
@@ -188,7 +190,8 @@ describe("codesyncd: recallDaemon", () => {
 
     test("with populateBuffer acquried by other instance", () => {
         lockFile.lockSync(settings.POPULATE_BUFFER_LOCK_FILE);
-        acquireSendDiffsLock();
+        const lockUtils = new LockUtils();
+        lockUtils.acquireSendDiffsLock();
         recallDaemon(statusBarItem);
         expect(CodeSyncState.get(CODESYNC_STATES.POPULATE_BUFFER_LOCK_ACQUIRED)).toBe(false);
 		expect(CodeSyncState.get(CODESYNC_STATES.DIFFS_SEND_LOCK_ACQUIRED)).toBe(true);